+++ example/config.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { createSimpleConfig } from "../src";
+import { createSimpleConfig, InputValidationError } from "../src";
 
 /**
  * @see onSubscribe
@@ -28,10 +28,17 @@ export const config = createSimpleConfig({
     rooms: {
       schema: z.tuple([z.string().array().describe("room IDs")]),
     },
+    error: {
+      schema: z.tuple([
+        z.string().describe("name"),
+        z.string().describe("message"),
+      ]),
+    },
   },
   examples: {
     time: { payload: ["2024-03-28T21:13:15.084Z"] },
     chat: { payload: ["Hello there!", { from: "123abc" }] },
+    error: { payload: ["InputValidationError", "1: Required"] },
     rooms: [
       { payload: [["room1", "room2"]] },
       { payload: [["room3", "room4", "room5"]] },
@@ -50,6 +57,14 @@ export const config = createSimpleConfig({
       setInterval(() => {
         withRooms(subscribersRoom).broadcast("time", new Date()); // <— payload type constraints
       }, 1000);
+    },
+    onError: async ({ error, client, logger, event }) => {
+      logger.error(event ? `${event} handling error` : "Error", error);
+      if (error instanceof InputValidationError && client) {
+        try {
+          await client.emit("error", error.name, error.message);
+        } catch {} // no errors inside this hook
+      }
     },
   },
   metadata: z.object({