+++ src/action.spec.ts
@@ -2,6 +2,7 @@ import { Socket } from "socket.io";
 import { describe, expect, test, vi } from "vitest";
 import { z } from "zod";
 import { AbstractAction, Action } from "./action";
+import { OutputValidationError, InputValidationError } from "./errors";
 import { AbstractLogger } from "./logger";
 
 describe("Action", () => {
@@ -59,150 +60,129 @@ describe("Action", () => {
       error: vi.fn(),
       debug: vi.fn(),
     };
-    const emitMock = vi.fn();
-    const broadcastMock = vi.fn();
-    const isConnectedMock = vi.fn();
-    const getRequestMock = vi.fn();
-    const withRoomsMock = vi.fn();
-    const getRoomsMock = vi.fn();
-    const getAllRoomsMock = vi.fn();
-    const getAllClientsMock = vi.fn();
-    const allBroadcastMock = vi.fn();
-    const getDataMock = vi.fn();
-    const setDataMock = vi.fn();
-    const joinMock = vi.fn();
-    const leaveMock = vi.fn();
+
+    const commons = {
+      withRooms: vi.fn(),
+      all: {
+        getClients: vi.fn(),
+        getRooms: vi.fn(),
+        broadcast: vi.fn(),
+      },
+      client: {
+        id: "ID",
+        handshake: { auth: {} } as Socket["handshake"],
+        emit: vi.fn(),
+        broadcast: vi.fn(),
+        getRooms: vi.fn(),
+        isConnected: vi.fn(),
+        getRequest: vi.fn(),
+        getData: vi.fn(),
+        setData: vi.fn(),
+        join: vi.fn(),
+        leave: vi.fn(),
+      },
+    };
 
     test("should handle simple action", async () => {
       await simpleAction.execute({
+        ...commons,
         event: "test",
         logger: loggerMock as unknown as AbstractLogger,
         params: ["some"],
-        withRooms: withRoomsMock,
-        all: {
-          getClients: getAllClientsMock,
-          getRooms: getAllRoomsMock,
-          broadcast: allBroadcastMock,
-        },
-        client: {
-          id: "ID",
-          handshake: { auth: {} } as Socket["handshake"],
-          emit: emitMock,
-          broadcast: broadcastMock,
-          getRooms: getRoomsMock,
-          isConnected: isConnectedMock,
-          getRequest: getRequestMock,
-          getData: getDataMock,
-          setData: setDataMock,
-          join: joinMock,
-          leave: leaveMock,
-        },
       });
       expect(loggerMock.error).not.toHaveBeenCalled();
       expect(simpleHandler).toHaveBeenLastCalledWith({
+        ...commons,
         input: ["some"],
         logger: loggerMock,
-        withRooms: withRoomsMock,
-        all: {
-          getClients: getAllClientsMock,
-          getRooms: getAllRoomsMock,
-          broadcast: allBroadcastMock,
-        },
-        client: {
-          id: "ID",
-          handshake: { auth: {} },
-          emit: emitMock,
-          broadcast: broadcastMock,
-          isConnected: isConnectedMock,
-          getRequest: getRequestMock,
-          getRooms: getRoomsMock,
-          getData: getDataMock,
-          setData: setDataMock,
-          join: joinMock,
-          leave: leaveMock,
-        },
       });
     });
 
     test("should handle action with ack", async () => {
       const ackMock = vi.fn();
       await ackAction.execute({
+        ...commons,
         event: "test",
         logger: loggerMock as unknown as AbstractLogger,
         params: ["some", ackMock],
-        withRooms: withRoomsMock,
-        all: {
-          getClients: getAllClientsMock,
-          getRooms: getAllRoomsMock,
-          broadcast: allBroadcastMock,
-        },
-        client: {
-          id: "ID",
-          handshake: { auth: {} } as Socket["handshake"],
-          emit: emitMock,
-          broadcast: broadcastMock,
-          getRooms: getRoomsMock,
-          isConnected: isConnectedMock,
-          getRequest: getRequestMock,
-          getData: getDataMock,
-          setData: setDataMock,
-          join: joinMock,
-          leave: leaveMock,
-        },
       });
       expect(loggerMock.error).not.toHaveBeenCalled();
       expect(ackHandler).toHaveBeenLastCalledWith({
-        client: {
-          id: "ID",
-          handshake: { auth: {} },
-          getRooms: getRoomsMock,
-          isConnected: isConnectedMock,
-          getRequest: getRequestMock,
-          emit: emitMock,
-          broadcast: broadcastMock,
-          getData: getDataMock,
-          setData: setDataMock,
-          join: joinMock,
-          leave: leaveMock,
-        },
-        all: {
-          getClients: getAllClientsMock,
-          getRooms: getAllRoomsMock,
-          broadcast: allBroadcastMock,
-        },
-        withRooms: withRoomsMock,
+        ...commons,
         input: ["some"],
         logger: loggerMock,
       });
       expect(ackMock).toHaveBeenLastCalledWith(123); // from ackHandler
     });
 
-    test("should catch errors", async () => {
-      await simpleAction.execute({
-        event: "test",
-        logger: loggerMock as unknown as AbstractLogger,
-        params: [], // too short
-        withRooms: withRoomsMock,
-        all: {
-          getClients: getAllClientsMock,
-          getRooms: getAllRoomsMock,
-          broadcast: allBroadcastMock,
-        },
-        client: {
-          id: "ID",
-          handshake: { auth: {} } as Socket["handshake"],
-          getRooms: getRoomsMock,
-          isConnected: isConnectedMock,
-          getRequest: getRequestMock,
-          emit: emitMock,
-          broadcast: broadcastMock,
-          getData: getDataMock,
-          setData: setDataMock,
-          join: joinMock,
-          leave: leaveMock,
-        },
-      });
-      expect(loggerMock.error).toHaveBeenCalled();
+    test("should throw input parsing error", async () => {
+      await expect(
+        simpleAction.execute({
+          ...commons,
+          event: "test",
+          logger: loggerMock as unknown as AbstractLogger,
+          params: [], // too short
+        }),
+      ).rejects.toThrowError(
+        new InputValidationError(
+          new z.ZodError([
+            {
+              code: "too_small",
+              minimum: 1,
+              inclusive: true,
+              exact: false,
+              type: "array",
+              path: [],
+              message: "Array must contain at least 1 element(s)",
+            },
+          ]),
+        ),
+      );
+    });
+
+    test.each([
+      [
+        "not cb",
+        new InputValidationError(
+          new z.ZodError([
+            {
+              code: "invalid_type",
+              expected: "function",
+              received: "string",
+              path: [1],
+              message: "Expected function, received string",
+            },
+          ]),
+        ),
+      ],
+      [
+        vi.fn(),
+        new OutputValidationError(
+          new z.ZodError([
+            {
+              code: "invalid_type",
+              expected: "number",
+              received: "string",
+              path: [0],
+              message: "Expected number, received string",
+            },
+          ]),
+        ),
+      ],
+    ])("should throw acknowledgment related errors %#", async (ack, error) => {
+      if (typeof ack === "function") {
+        ackHandler.mockImplementationOnce(async () => [
+          "not number" as unknown as number,
+        ]);
+      }
+      await expect(
+        ackAction.execute({
+          ...commons,
+          event: "test",
+          logger: loggerMock as unknown as AbstractLogger,
+          params: ["test", ack],
+        }),
+      ).rejects.toThrowError(error);
     });
   });
 });