+++ src/common-helpers.ts
@@ -38,3 +38,17 @@ export const makeCleanId = (...args: string[]) =>
     )
     .map(ucFirst)
     .join("");
+
+export const makeErrorFromAnything = (subject: unknown): Error =>
+  subject instanceof Error ? subject : new Error(String(subject));
+
+export const getMessageFromError = (error: Error): string => {
+  if (error instanceof z.ZodError) {
+    return error.issues
+      .map(({ path, message }) =>
+        (path.length ? [path.join("/")] : []).concat(message).join(": "),
+      )
+      .join("; ");
+  }
+  return error.message;
+};