+++ vitest.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
     },
     pool: "threads",
     testTimeout: 10000,
-    reporters: "verbose",
+    reporters: "basic",
     coverage: {
       provider: "istanbul",
       reporter: [["text", { maxCols: 120 }], "json-summary", "html", "lcov"],