+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 ## Version 2
 
+### v2.1.0
+
+- Featuring `onError` hook for handling errors of various natures:
+  - The hook is intended to be generic, so some of its arguments are optional;
+  - Exposing error classes: `InputValidationError` and `OutputValidationError` (for Action acknowledgments);
+  - The following example shows how to emit an outgoing `error` event when the incoming event data is invalid:
+
+```typescript
+import { createSimpleConfig, InputValidationError } from "zod-sockets";
+
+const config = createSimpleConfig({
+  emission: {
+    error: {
+      schema: z.tuple([
+        z.string().describe("name"),
+        z.string().describe("message"),
+      ]),
+    },
+  },
+  hooks: {
+    onError: async ({ error, event, payload, client, logger }) => {
+      logger.error(event ? `${event} handling error` : "Error", error);
+      if (error instanceof InputValidationError && client) {
+        try {
+          await client.emit("error", error.name, error.message);
+        } catch {} // no errors inside this hook
+      }
+    },
+  },
+});
+```
+
 ### v2.0.1
 
 - Technical update due to improved builder configuration: